+++ packages/flame/lib/src/game/base_game.dart
@@ -11,6 +11,7 @@ import '../components/mixins/collidable.dart';
 import '../components/mixins/draggable.dart';
 import '../components/mixins/has_collidables.dart';
 import '../components/mixins/has_game_ref.dart';
+import '../components/mixins/hoverable.dart';
 import '../components/mixins/tapable.dart';
 import '../components/position_component.dart';
 import '../fps_counter.dart';
@@ -104,6 +105,12 @@ class BaseGame extends Game with FPSCounter {
       assert(
         this is HasDraggableComponents,
         'Draggable Components can only be added to a BaseGame with HasDraggableComponents',
+      );
+    }
+    if (c is Hoverable) {
+      assert(
+        this is HasHoverableComponents,
+        'Hoverable Components can only be added to a BaseGame with HasHoverableComponents',
       );
     }
 
@@ -237,6 +244,58 @@ class BaseGame extends Game with FPSCounter {
   /// You can use this value to enable debug behaviors for your game and many components will
   /// show extra information on the screen when debug mode is activated
   bool debugMode = false;
+
+  /// Changes the priority of [component] and reorders the games component list.
+  ///
+  /// Returns true if changing the component's priority modified one of the
+  /// components that existed directly on the game and false if it
+  /// either was a child of another component, if it didn't exist at all or if
+  /// it was a component added directly on the game but its priority didn't
+  /// change.
+  bool changePriority(
+    Component component,
+    int priority, {
+    bool reorderRoot = true,
+  }) {
+    if (component.priority == priority) {
+      return false;
+    }
+    component.changePriorityWithoutResorting(priority);
+    if (reorderRoot) {
+      if (component.parent != null && component.parent is BaseComponent) {
+        (component.parent! as BaseComponent).reorderChildren();
+      } else if (components.contains(component)) {
+        components.rebalanceAll();
+      }
+    }
+    return true;
+  }
+
+  /// Since changing priorities is quite an expensive operation you should use
+  /// this method if you want to change multiple priorities at once so that the
+  /// tree doesn't have to be reordered multiple times.
+  void changePriorities(Map<Component, int> priorities) {
+    var hasRootComponents = false;
+    final parents = <BaseComponent>{};
+    priorities.forEach((component, priority) {
+      final wasUpdated = changePriority(
+        component,
+        priority,
+        reorderRoot: false,
+      );
+      if (wasUpdated) {
+        if (component.parent != null && component.parent is BaseComponent) {
+          parents.add(component.parent! as BaseComponent);
+        } else {
+          hasRootComponents |= components.contains(component);
+        }
+      }
+    });
+    if (hasRootComponents) {
+      components.rebalanceAll();
+    }
+    parents.forEach((parent) => parent.reorderChildren());
+  }
 
   /// Returns the current time in seconds with microseconds precision.
   ///